# CUDA: Prefer vector flash decoding kernel for Gemma models